Dev - Create pull request

+++ ecommerce-backend/src/main/java/com/nashtech/rookies/ecommerce/services/cart/impl/CartItemServiceImpl.java
@@ -3,13 +3,19 @@ package com.nashtech.rookies.ecommerce.services.cart.impl;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.nashtech.rookies.ecommerce.repositories.cart.OrderRepository;
+import com.nashtech.rookies.ecommerce.dto.cart.responses.PaginationCartItemDTO;
+import com.nashtech.rookies.ecommerce.models.cart.Cart;
+import com.nashtech.rookies.ecommerce.models.prod.Product;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.nashtech.rookies.ecommerce.dto.cart.requests.CartItemRequestDTO;
 import com.nashtech.rookies.ecommerce.dto.cart.responses.CartItemResponseDTO;
-import com.nashtech.rookies.ecommerce.exceptions.ResourceNotFoundException;
+import com.nashtech.rookies.ecommerce.handlers.exceptions.NotFoundException;
 import com.nashtech.rookies.ecommerce.mappers.cart.CartItemMapper;
 import com.nashtech.rookies.ecommerce.models.cart.CartItem;
 import com.nashtech.rookies.ecommerce.repositories.cart.CartItemRepository;
@@ -40,39 +46,50 @@ public class CartItemServiceImpl extends CommonServiceImpl<CartItem, Long> imple
     public CartItemResponseDTO createCartItem(CartItemRequestDTO cartItemRequestDTO) {
         CartItem cartItem = new CartItem();
         if (cartRepository.existsById(cartItemRequestDTO.cartId())) {
-            cartItem.setCart(cartRepository.findById(cartItemRequestDTO.cartId()).get());
+            Cart cart = cartRepository.findById(cartItemRequestDTO.cartId()).get();
+            cartItem.setCart(cart);
             if (productRepository.existsById(cartItemRequestDTO.productId())) {
                 cartItem.setProduct(productRepository.findById(cartItemRequestDTO.productId()).get());
                 cartItem.setQuantity(cartItemRequestDTO.quantity());
                 cartItem = cartItemRepository.saveAndFlush(cartItem);
+                cart.setQuantity(cart.getQuantity() + 1L);
+                cartRepository.saveAndFlush(cart);
                 return cartItemMapper.toResponseDTO(cartItem);
             } else
-                throw new ResourceNotFoundException("Not found Product with an id: " + cartItemRequestDTO.productId());
+                throw new NotFoundException("Not found Product with an id: " + cartItemRequestDTO.productId());
         } else
-            throw new ResourceNotFoundException("Not found Cart with an id: " + cartItemRequestDTO.cartId());
+            throw new NotFoundException("Not found Cart with an id: " + cartItemRequestDTO.cartId());
     }
 
-//    new CartItemResponseDTO(cartItem.getId(), cartItem.getCreatedOn(), cartItem.getLastUpdatedOn(),
-//                            cartItem.getQuantity(), cartItem.getCart().getId(), cartItem.getOrder().getId(),
-//                            cartItem.getProduct().getId())
-
     @Override
-    public List<CartItemResponseDTO> getCartItem() {
-        var cartItems = cartItemRepository.findAll();
+    public PaginationCartItemDTO getCartItem(Sort.Direction dir, int pageNum, int pageSize) {
+        Sort sort = Sort.by(dir, "id");
+        Pageable pageable = PageRequest.of(pageNum, pageSize, sort);
+        Page<CartItem> cartItems = cartItemRepository.findAll(pageable);
         List<CartItemResponseDTO> cartItemResponseDTOs = new ArrayList<>();
         cartItems.forEach(cartItem -> cartItemResponseDTOs.add(cartItemMapper.toResponseDTO(cartItem)));
-        return cartItemResponseDTOs;
+        return new PaginationCartItemDTO(cartItems.getTotalPages(), cartItems.getTotalElements(), cartItems.getSize(),
+                cartItems.getNumber(), cartItemResponseDTOs);
     }
 
     @Override
-    public List<CartItemResponseDTO> getCartItem(Long id) {
+    public PaginationCartItemDTO getCartItemByUserId(Long userId, Sort.Direction dir, int pageNum, int pageSize) {
+        Sort sort = Sort.by(dir, "id");
+        Pageable pageable = PageRequest.of(pageNum, pageSize, sort);
+        Page<CartItem> cartItems = cartItemRepository.findByUserId(userId, pageable);
+        List<CartItemResponseDTO> cartItemResponseDTOs = new ArrayList<>();
+        cartItems.forEach(cartItem -> cartItemResponseDTOs.add(cartItemMapper.toResponseDTO(cartItem)));
+        return new PaginationCartItemDTO(cartItems.getTotalPages(), cartItems.getTotalElements(), cartItems.getSize(),
+                cartItems.getNumber(), cartItemResponseDTOs);
+    }
+
+    @Override
+    public CartItemResponseDTO getCartItem(Long id) {
         if (cartItemRepository.existsById(id)) {
             CartItem cartItem = cartItemRepository.findById(id).get();
-            List<CartItemResponseDTO> cartItemResponseDTOs = new ArrayList<>();
-            cartItemResponseDTOs.add(cartItemMapper.toResponseDTO(cartItem));
-            return cartItemResponseDTOs;
+            return cartItemMapper.toResponseDTO(cartItem);
         } else
-            throw new ResourceNotFoundException("Not found a Cart Item with an id: " + id);
+            throw new NotFoundException("Not found a Cart Item with an id: " + id);
     }
 
     @Override
@@ -83,13 +100,21 @@ public class CartItemServiceImpl extends CommonServiceImpl<CartItem, Long> imple
             cartItem = cartItemRepository.saveAndFlush(cartItem);
             return cartItemMapper.toResponseDTO(cartItem);
         } else
-            throw new ResourceNotFoundException("Not found a Cart Item with an id: " + id);
+            throw new NotFoundException("Not found a Cart Item with an id: " + id);
     }
 
     @Override
     @Transactional
     public String deleteCartItem(Long id) {
-        cartItemRepository.deleteById(id);
-        return "Delete Cart Item successful";
+        if (cartItemRepository.existsById(id)) {
+            CartItem cartItem = cartItemRepository.findById(id).get();
+            Cart cart = cartRepository.findById(cartItem.getCart().getId()).get();
+            cart.setQuantity(cart.getQuantity() - 1);
+            cartRepository.saveAndFlush(cart);
+            cartItemRepository.deleteById(id);
+            return "Delete successful";
+        } else {
+            throw new NotFoundException("Not found a Cart Item with an id: " + id);
+        }
     }
 }
